test_folder/requirements.txt
python-dotenv
plotly
langchain
langgraph
langchain-groq
sqlalchemy
streamlit
mysql-connector-python
langchain-community